ci: add a native musl libc build to the test matrix
Add an x86_64 build-check matrix entry that runs inside the GNU-userland
musl rootfs image published at ghcr.io/gabifalk/strace-musl-image, so
configure's run-time feature checks execute against musl rather than
being guessed and the resulting binaries are exercised natively.

* .github/workflows/ci.yml (build-check): Add an x86_64 musl matrix
entry carrying a container image, run the job in matrix.container, select
--disable-stacktrace for the musl variant, and skip the apt dependency
step when running in a container.

+++ tests/sync_file_range.c
@@ -20,8 +20,8 @@ int
 main(void)
 {
 	const int fd = -1;
-	const off64_t offset = 0xdeadbeefbadc0dedULL;
-	const off64_t nbytes = 0xfacefeedcafef00dULL;
+	const long long offset = 0xdeadbeefbadc0dedULL;
+	const long long nbytes = 0xfacefeedcafef00dULL;
 	const unsigned int flags = -1;
 
 	int rc = sync_file_range(fd, offset, nbytes, flags);
